[#7] refactor: 이름, 배송상태 쿼리스트링 구현
