+++ src/components/index.ts
@@ -1,4 +1,5 @@
 export { default as MainTable } from './table/MainTable';
 export { default as PagenationBar } from './tools/PaginationBar';
-export { default as HeadersFilters } from './tools/columns/HeadersFilters';
 export { default as HeadersSort } from './tools/columns/HeadersSort';
+export { default as NameSearch } from './tools/columns/NameSearch';
+export { default as StatusFilter } from './tools/columns/StatusFilter';
