+++ src/pages/MainPage.tsx
@@ -1,13 +1,12 @@
 import { createContext, useEffect, useState } from 'react';
 
-import { useQuery } from '@tanstack/react-query';
-
 import { PagenationBar } from 'components';
 import LayoutWrapper from 'components/layout/LayoutWrapper';
 import MainTable from 'components/table/MainTable';
 import InfoContainer from 'components/tools/InfoContainer';
 import useColums from 'hooks/useColums';
 import useOrderList from 'hooks/useOrderList';
+import useParamsFilter from 'hooks/useParamsFilter';
 import useTableOptions from 'hooks/useTableOptions';
 import { FILTER_DATE } from 'types/constans';
 
@@ -19,12 +18,18 @@ const MainPage = () => {
   const [isFilterReset, setIsFilterRest] = useState(false);
   const columns = useColums();
   const table = useTableOptions({ data, columns });
+  const { filterState, reSeatStatae } = useParamsFilter();
 
   const onResetFilterUIHandler = () => {
     table.setColumnFilters([initialFilter]);
+    reSeatStatae();
     setIsFilterRest(true);
     setTimeout(() => setIsFilterRest(false), 1000);
   };
+
+  useEffect(() => {
+    table.setColumnFilters(filterState);
+  }, []);
 
   return (
     <LayoutWrapper>
